Detabified and fixed the code, we need to debug now and I am adding the save and load code

+++ src/game_of_life.py
@@ -1,6 +1,6 @@
 __author__ = "scrum-diddlyumptious"
 
-from Tkinter import *
+from tkinter import *
 import pandas as pd
 import time
 
@@ -12,8 +12,8 @@ storedGrid = []
 storedGridIndex = 0
 windowCanvasWidth = 500
 windowCanvasHeight = 500
-cellWidth = windowCanvasWidth / currentRows
-cellHeight = windowCanvasHeight / currentColumns
+cellWidth = int(windowCanvasWidth / currentRows)
+cellHeight = int(windowCanvasHeight / currentColumns)
 stepNum = 0
 lifeNum = 0
 running = False
@@ -36,145 +36,143 @@ waitTime = 1.0 #start with 1 second slowdown of game run speed(later can multipl
 def windows_menu(root,windowCanvas, e):
 
 	# Reads the file
-	def command_read_file():
-	   fileWin = Toplevel(root)
-	   button = Button(fileWin, text="Ryan will implement read file function")
-	   button.pack()
+        def command_read_file():
+                fileWin = Toplevel(root)
+                button = Button(fileWin, text="Ryan will implement read file function")
+                button.pack()
 
 	# Write to file
-	def command_write_file():
-	   fileWin = Toplevel(root)
-	   button = Button(fileWin, text="Ryan will implement write file function")
-	   button.pack()
+        def command_write_file():
+                fileWin = Toplevel(root)
+                button = Button(fileWin, text="Ryan will implement write file function")
+                button.pack()
 
    # Skeleton code for menu buttons actions
-	def do_nothing():
-		fileWin = Toplevel(root)
-		button = Button(fileWin, text="Do nothing button")
-		button.pack()
+        def do_nothing():
+                fileWin = Toplevel(root)
+                button = Button(fileWin, text="Do nothing button")
+                button.pack()
 
 	# change the grid color
-	def change_color(changedColors):
-		global currentGridColor
-		currentGridColor = changedColors
+        def change_color(changedColors):
+                global currentGridColor
+                currentGridColor = changedColors
 
 	# creates the grid line
-	def create_grid(windowCanvas):
+        def create_grid(windowCanvas):
 		# draw horizontal lines
-		x1 = 0
-		x2 = windowCanvasWidth+1
-		for k in range(0, currentRows):
-			y1 = k*cellHeight
-			y2 = k*cellHeight
-			windowCanvas.create_line(x1, y1, x2, y2, tag='grid_line')
+                x1 = 0
+                x2 = windowCanvasWidth+1
+                for k in range(0, currentRows):
+                        y1 = k*cellHeight
+                        y2 = k*cellHeight
+                        windowCanvas.create_line(x1, y1, x2, y2, tag='grid_line')
 
 		# draw vertical lines
-		y1 = 0
-		y2 = windowCanvasHeight+1
-		for k in range(0, currentColumns):
-			x1 = k*cellWidth
-			x2 = k*cellWidth
-			windowCanvas.create_line(x1, y1, x2, y2, tag='grid_line')
+                y1 = 0
+                y2 = windowCanvasHeight+1
+                for k in range(0, currentColumns):
+                        x1 = k*cellWidth
+                        x2 = k*cellWidth
+                        windowCanvas.create_line(x1, y1, x2, y2, tag='grid_line')
 
 	
 	
 	# initialize parent menus
-	menuBar = Menu(root)
-	fileMenu = Menu(menuBar, tearoff=0)
-	showMenu = Menu(menuBar, tearoff=0)
-	actionsMenu = Menu(menuBar, tearoff=0)
-	speedMenu = Menu(menuBar, tearoff=0)
-	subMenuColor = Menu(menuBar, tearoff=0)
+        menuBar = Menu(root)
+        fileMenu = Menu(menuBar, tearoff=0)
+        showMenu = Menu(menuBar, tearoff=0)
+        actionsMenu = Menu(menuBar, tearoff=0)
+        speedMenu = Menu(menuBar, tearoff=0)
+        subMenuColor = Menu(menuBar, tearoff=0)
 
 	# first drop down menu "FILE"
-	menuBar.add_cascade(label="File", menu=fileMenu)
-	fileMenu.add_command(label="Read File", command=command_read_file)
-	fileMenu.add_command(label="Write File", command=command_write_file)
+        menuBar.add_cascade(label="File", menu=fileMenu)
+        fileMenu.add_command(label="Read File", command=command_read_file)
+        fileMenu.add_command(label="Write File", command=command_write_file)
 
 	# second drop down menu "Show"
-	menuBar.add_cascade(label="Show", menu=showMenu)
-	showMenu.add_command(label="Static", command=do_nothing)
-	showMenu.add_command(label="Oscillate", command=do_nothing)
-	showMenu.add_command(label="Move", command=do_nothing)
-	showMenu.add_command(label="Weird", command=do_nothing)
+        menuBar.add_cascade(label="Show", menu=showMenu)
+        showMenu.add_command(label="Static", command=do_nothing)
+        showMenu.add_command(label="Oscillate", command=do_nothing)
+        showMenu.add_command(label="Move", command=do_nothing)
+        showMenu.add_command(label="Weird", command=do_nothing)
 
 	# third drop down menu "Actions"
-	menuBar.add_cascade(label="Actions", menu=actionsMenu)
-	actionsMenu.add_command(label="Clear", command=do_nothing)
-	actionsMenu.add_command(label="Run", command=lambda:run(df, windowCanvas, root, e))
-	actionsMenu.add_command(label="Step", command=lambda:refresh_life(df, windowCanvas, root))
-	actionsMenu.add_command(label="Stop", command=do_nothing)
-	actionsMenu.add_command(label="Quit", command=root.quit)
+        menuBar.add_cascade(label="Actions", menu=actionsMenu)
+        actionsMenu.add_command(label="Clear", command=do_nothing)
+        actionsMenu.add_command(label="Run", command=lambda:run(df, windowCanvas, root, e))
+        actionsMenu.add_command(label="Step", command=lambda:refresh_life(df, windowCanvas, root))
+        actionsMenu.add_command(label="Stop", command=do_nothing)
+        actionsMenu.add_command(label="Quit", command=root.quit)
 
 	# fourth drop down menu "Grid/Speed"
-	menuBar.add_cascade(label="Grid/Speed", menu=speedMenu)
-	speedMenu.add_command(label="No Grid", command=lambda:windowCanvas.delete('grid_line'))
-	speedMenu.add_command(label="Show Grid", command=lambda: create_grid(windowCanvas))
-	speedMenu.add_command(label="Faster", command=do_nothing)
-	speedMenu.add_command(label="Slower", command=do_nothing)
-	speedMenu.add_cascade(label="Color", command=do_nothing, menu=subMenuColor)
+        menuBar.add_cascade(label="Grid/Speed", menu=speedMenu)
+        speedMenu.add_command(label="No Grid", command=lambda:windowCanvas.delete('grid_line'))
+        speedMenu.add_command(label="Show Grid", command=lambda: create_grid(windowCanvas))
+        speedMenu.add_command(label="Faster", command=do_nothing)
+        speedMenu.add_command(label="Slower", command=do_nothing)
+        speedMenu.add_cascade(label="Color", command=do_nothing, menu=subMenuColor)
 
 	# calling submenu color functions
-	subMenuColor.add_command(label="red",command=lambda:change_color("red"))
-	subMenuColor.add_command(label="blue",command=lambda:change_color("blue"))
-	subMenuColor.add_command(label="green",command=lambda:change_color("green"))
+        subMenuColor.add_command(label="red",command=lambda:change_color("red"))
+        subMenuColor.add_command(label="blue",command=lambda:change_color("blue"))
+        subMenuColor.add_command(label="green",command=lambda:change_color("green"))
 
 	# Actually register and display the four menu dropdown menu
-	root.config(menu=menuBar)
+        root.config(menu=menuBar)
 
 def step_counter(root):
-       	global stepNum
-       	stepNum+=1
-       	stepCounterLabel = Label(root, text=stepNum)
-       	stepCounterLabel.place(x=85, y=1)
+        global stepNum
+        stepNum+=1
+        stepCounterLabel = Label(root, text=stepNum)
+        stepCounterLabel.place(x=85, y=1)
 
 def life_counter(root):
-	global lifeNum
+        global lifeNum
         global df
-	lifeNum = len(df['pos'].tolist())
-	lifeCounterLabel = Label(root, text=lifeNum)
-	lifeCounterLabel.place(x=windowCanvasWidth-35, y=1)            
-
+        lifeNum = len(df['pos'].tolist())
+        lifeCounterLabel = Label(root, text=lifeNum)
+        lifeCounterLabel.place(x=windowCanvasWidth-35, y=1)             
 # Litterly copied from this website:
 # https://stackoverflow.com/questions/26988204/using-2d-array-to-create-clickable-tkinter-canvas
 def clickable_grid(root,windowCanvas):
 
 	# Create a grid of None to store the references to the tiles
-	global df
+        global df
 
 	# Asynchronous listener
 	# This is acutally the backend of the clickable grid
-	def callback(event):
+        def callback(event):
 
 		# Storing and retrieving the grid length
-		global storedGridIndex
+                global storedGridIndex
 
 		# Get rectangle diameters
-		columnWidth = cellWidth
-		rowHeight = cellHeight
+                columnWidth = cellWidth
+                rowHeight = cellHeight
 
 		# Calculate column and row number
-		columnNum = event.x//columnWidth
-		rowNum = event.y//rowHeight
+                columnNum = event.x//columnWidth
+                rowNum = event.y//rowHeight
 
 	
 		# If the tile is not filled, create a rectangle
 		# Also sets the grid color as well
-		if not tiles[rowNum][columnNum]:
-			tiles[rowNum][columnNum] = windowCanvas.create_rectangle(columnNum*columnWidth, rowNum*rowHeight, (columnNum+1)*columnWidth, (rowNum+1)*rowHeight, fill=currentGridColor,outline=currentGridColor)
+                if not tiles[rowNum][columnNum]:
+                        tiles[rowNum][columnNum] = windowCanvas.create_rectangle(columnNum*columnWidth, rowNum*rowHeight, (columnNum+1)*columnWidth, (rowNum+1)*rowHeight, fill=currentGridColor,outline=currentGridColor)
                         storedGrid.append([])
-		        storedGrid[storedGridIndex].append(columnNum)
-		        storedGrid[storedGridIndex].append(rowNum)
+                        storedGrid[storedGridIndex].append(columnNum)
+                        storedGrid[storedGridIndex].append(rowNum)
                         print('index: ',storedGridIndex)
                         df.loc[storedGridIndex] = [currentGridColor, [columnNum, rowNum]]
-		
-		        storedGridIndex+=1
+                        storedGridIndex+=1
 
 		# If the tile is filled, delete the rectangle and clear the reference
-		else:
+                else:
                         drop = df['pos'].tolist().index([columnNum, rowNum])
-			windowCanvas.delete(tiles[rowNum][columnNum])
-			tiles[rowNum][columnNum] = None
+                        windowCanvas.delete(tiles[rowNum][columnNum])
+                        tiles[rowNum][columnNum] = None
                         del storedGrid[storedGrid.index([columnNum, rowNum])]
                         df.drop(index=drop, inplace=True)
                         storedGridIndex-=1
@@ -184,8 +182,8 @@ def clickable_grid(root,windowCanvas):
                 print(storedGrid) # prints the coordinates on the terminal/output
                 #print(df)
 	# figures out how the canvas sits in the window
-	windowCanvas.pack()
-	windowCanvas.bind("<Button-1>", callback)
+        windowCanvas.pack()
+        windowCanvas.bind("<Button-1>", callback)
 
         
 def run(df, windowCanvas, root, e):
@@ -196,23 +194,23 @@ def run(df, windowCanvas, root, e):
                 runLoop(df, windowCanvas, root, e, iterationCount) #triggers running loop
 
 def runLoop(df, windowCanvas, root, e, iterationCount): #added this so run function can enter an endless loop that only exits when global variable is modified
-    if(running):
+        if(running):
                 refresh_life(df, windowCanvas, root)
                 if(iterationCount == FIRST_ITERATION):
-                    df2Prev = df #first df to record is for 2 iterations ago
-                    iterationCount += 1
+                        df2Prev = df #first df to record is for 2 iterations ago
+                        iterationCount += 1
                 elif(iterationCount == SECOND_ITERATION):
-                    dfPrev = df #now record the 'most recent' df iteration second
-                    iterationCount += 1
+                        dfPrev = df #now record the 'most recent' df iteration second
+                        iterationCount += 1
                 elif(iterationCount == STABILITY_CHECK_ITERATION):
                     #now its time to check for stability, then reset counter
-                    if(check_stable(df, df2Prev)):
+                        if(check_stable(df, df2Prev)):
                         #(insert code here to set GUI's stable notification to visible)
-                        print("(insert code here to set GUI's stable notification to visible)")
-                    else:
+                                print("(insert code here to set GUI's stable notification to visible)")
+                        else:
                         #(and insert code here to set GUI's stable notification to invisible)
-                        print("(insert code here to set GUI's stable notification to invisible)")
-                    iterationCount = FIRST_ITERATION #reset counter
+                                print("(insert code here to set GUI's stable notification to invisible)")
+                iterationCount = FIRST_ITERATION #reset counter
                 root.update()
                 root.after(waitTime, lambda:runLoop(df, windowCanvas, root, e, iterationCount)) #recursively calls itself
 
@@ -323,25 +321,23 @@ def refresh_life(df, windowCanvas, root):
         for i in range(0, currentColumns):
                 for j in range(0, currentRows):
                         # Get rectangle diameters
-	                columnWidth = cellWidth
-	                rowHeight = cellHeight
+                        columnWidth = cellWidth
+                        rowHeight = cellHeight
                         
 	                # Calculate column and row number
-	                columnNum = i
-	                rowNum = j
-
+                        columnNum = i
+                        rowNum = j
                         currentGridColor = 'white' if [i,j] not in df['pos'].tolist() else df['color'].tolist()[df['pos'].tolist().index([i,j])]
-        
                                         
 		        # If the tile is not filled, create a rectangle
 		        # Also sets the grid color as well
-		        if not tiles[rowNum][columnNum] and [i,j] in nextGrid:
-			        tiles[rowNum][columnNum] = windowCanvas.create_rectangle(columnNum*columnWidth, rowNum*rowHeight, (columnNum+1)*columnWidth, (rowNum+1)*rowHeight, fill=currentGridColor,outline = currentGridColor)
+                        if not tiles[rowNum][columnNum] and [i,j] in nextGrid:
+                                tiles[rowNum][columnNum] = windowCanvas.create_rectangle(columnNum*columnWidth, rowNum*rowHeight, (columnNum+1)*columnWidth, (rowNum+1)*rowHeight, fill=currentGridColor,outline = currentGridColor)
 
 		                # If the tile is filled, delete the rectangle and clear the reference
-		        if (tiles[rowNum][columnNum] != None and [i,j] not in nextGrid):
+                        if (tiles[rowNum][columnNum] != None and [i,j] not in nextGrid):
                                 windowCanvas.delete(tiles[rowNum][columnNum])
-			        tiles[rowNum][columnNum] = None
+                                tiles[rowNum][columnNum] = None
                                 
         life_counter(root)
 
@@ -350,36 +346,35 @@ def refresh_life(df, windowCanvas, root):
 def main():
 
 	# class alias to variable
-	root = Tk()
-
+        root = Tk()
 	# title of the program
-	root.title("Game of Life")
+        root.title("Game of Life")
         e = Entry(root)
 	# initialize window size and color
-	windowCanvas = Canvas(root, width=windowCanvasWidth, height=windowCanvasHeight, borderwidth=0, background='white', highlightbackground = 'black')    
+        windowCanvas = Canvas(root, width=windowCanvasWidth, height=windowCanvasHeight, borderwidth=0, background='white', highlightbackground = 'black')    
 
 	# create instructions
-	Label(root, text="Instructions:\nChoose From Dropdown Menu Or Left Click Grid Points \n Right click to run one step").pack()
+        Label(root, text="Instructions:\nChoose From Dropdown Menu Or Left Click Grid Points \n Right click to run one step").pack()
 
 	#Displaying step counter
-	Label(root, text="Step Count: ").place(x=0, y=1)
+        Label(root, text="Step Count: ").place(x=0, y=1)
 
 	#Displaying life counter
-	Label(root, text="Life Count: ").place(x=windowCanvasWidth-120, y=1)
+        Label(root, text="Life Count: ").place(x=windowCanvasWidth-120, y=1)
 
 	# backend grid listener
-	clickable_grid(root,windowCanvas)
+        clickable_grid(root,windowCanvas)
 
 	# creates the menu
-	windows_menu(root,windowCanvas, e)
+        windows_menu(root,windowCanvas, e)
 
 	# the pack geometry manager organises widgets in horizontal and vertical boxes
-	windowCanvas.pack()
+        windowCanvas.pack()
 
 	# start monitoring and updating the GUI.
-	root.mainloop()
+        root.mainloop()
 
 
 # Calls main function
 if __name__== "__main__":
-	main()
+        main()
